src/requirements.txt
PyYAML
google-cloud-bigquery
google-cloud-aiplatform
feedparser
tweepy
google-generativeai
requests
pydantic
Pillow
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
